requirements.txt
# Core
fastapi==0.110.1
uvicorn[standard]==0.29.0

# Async DB
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0

# Security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP Client for LLaMA
httpx==0.27.0
requests

# Redis Cache
aioredis==2.0.1

# Pydantic
pydantic[email]==2.6.4
email-validator==2.1.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5

 
